seperated stuff into individual files. much better organized now
