+++ server/src/apis/crop_image_handler.py
@@ -1,15 +1,12 @@
 from flask import jsonify
-from flask_restplus import Resource
+from flask_restplus import Namespace, Resource, fields
 
 DEFAULT_ID = -1 # default to this if no id is specified in the request
 
-class RawImageHandler(Resource):
-    def get(self, id=DEFAULT_ID):
-        if id == DEFAULT_ID:
-            return jsonify({"yousent": "nothin"})
-        else:
-            return jsonify({'yousent': id})
+api  = Namespace('image/crop', description="All imaging related calls route through here")
 
+@api.route('/')
+@api.route('/<int:id>')
 class CroppedImageHandler(Resource):
     def get(self, id=DEFAULT_ID):
         if id == DEFAULT_ID:
